requirements.txt
requests>=2.28.1
Pillow>=9.3.0 